// src/app/models/assignServiceDetail.ts
export class AssignServiceDetail {
    assignServiceDetailId: number;
    assignServiceId: number;
    professionalId: string;
    professionalName: string;
    dateVisit: string;
    consecutive: number;
    stateId: number;
    stateName: string;
    paymentName: string;
    paymentType: number;
    receivedAmount: number;
    otherAmount: number;
    observation: string;
    totalRows: number;
    pin: number;
    verified: boolean;
    verifiedBy: number;
    isQualityTestDone: boolean;
}